Added settings screen and modifiable position
I implemented a simple settings screen in
MPDMC.GUI.Screens.MPDOptionsScreen but the only problem with it is
that the background is not transparent. For this reason, the class
needs to extend Screen instead of GameOptionsScreen. I didn't do it
here since it never came to mind that the background will not be
transparent in-game. As for the movable text and settings, they are
found in the MPDMC.Settings class. Adding new settings is easy but not
recommended because of the problems listed above. We also need to
implement more settings which I didn;t have the time to implement.